% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sobolnp.R
\name{plot}
\alias{plot}
\alias{plot.sobolnp}
\title{Plot function for class \code{sobolnp}}
\usage{
plot(x, ...)

\method{plot}{sobolnp}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sobolnp}}

\item{...}{further arguments passed to the \code{plot} function}
}
\value{
Plot with Sobol indices
}
\description{
Plot the Sobol indices based in a non-parametric regression
with cross-validation and bootstrap bandwidth
}
\examples{

\dontrun{
ishigami.fun <- function(X) {
A <- 7
B <- 0.1
sin(X[, 1]) + A * sin(X[, 2])^2 + B * X[, 3]^4 * sin(X[, 1])
}

X<- matrix(runif(3*10, -pi,pi), ncol = 3)
Y<- ishigami.fun(X)

estimation <- sobolnp::sobolnp(Y = Y,X = X, mc.cores = 1)

plot(estimation)
}

}
