% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sobolnp.R
\name{print}
\alias{print}
\alias{print.sobolnp}
\title{Title}
\usage{
print(x, ...)

\method{print}{sobolnp}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sobolnp}}

\item{...}{further arguments passed to the \code{print} function}
}
\value{
A formatted
}
\description{
Title
}
\examples{

\dontrun{
ishigami.fun <- function(X) {
A <- 7
B <- 0.1
sin(X[, 1]) + A * sin(X[, 2])^2 + B * X[, 3]^4 * sin(X[, 1])
}

X<- matrix(runif(3*10, -pi,pi), ncol = 3)
Y<- ishigami.fun(X)

estimation <- sobolnp::sobolnp(Y = Y,X = X, mc.cores = 1)

print(estimation)
}

}
